Change Yahoo Finance download logic
Previously we were fetching FX data from Yahoo Finance for USD/CCY currency
pair with `CCY=X` ticker, the symbol CCY denotes the domestic currency. For
some currencies, however, this is not possible because the data for the ticker
`CCY=X` is unavailable. Alternatively, using the ticker `CCYUSD=X` provides
more data for a larger number of currencies and therefore we reimplement the
download logic to get the data for the currency pair USD/CCY. After we download
the data for CCY/USD we invert the currency pair and transform it to USD/CCY.